Allow client custom fields to be used in invoice group identifiers.
My preferred invoice format is ABC-123, where ABC is a code specific to
the client and 123 is the incrementing ID for that client. To do this, I
had to make a new invoice group for each client and statically set that
client's code in the identifier. This change allows me to set the client
code using a custom field and use a single invoice group for all my
clients.

An example format string for this is {{{client:Invoice prefix}}} where
"Invoice prefix" is the label of the custom field.

This also works for quotes.