Detect when credit bill
Cuando se carga la plantilla para la factura preguntamos si es de crédito o no. En base a la respuesta cargamos una plantilla u otra.

---

When loading the template for the invoice we ask if it is credit or not. Based on the answer we load a template or another.