[IP-742] Improve display of second street address line
In the email-template form, in the list of tags the fields
ip_clients.client_address_1 and ip_clients.client_address_2,
and  ip_users.user_address_1 and ip_users.user_address_2
are displayed with same tag "Address" because the code

<?php _trans('address') . ' 1'; ?>

fails to add the "1" or "2".

Fix: use 'street_address' and 'street_address_2' instead of 'address'

Additionally, the fix replaces "Street Address 2" with
"Street Address (cont.)" in the English language file.